+++ frontend/src/api/message.api.ts
@@ -1,0 +1,27 @@
+import { io } from "socket.io-client";
+
+export const socket = io("http://localhost:8001", {
+	transportOptions: {
+		polling: {
+			extraHeaders: {
+				"Authorization": `Bearer ${localStorage.getItem("token")}`
+			}
+		}
+	}
+});
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+